requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
whitenoise>=6.6
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
gunicorn>=21.0
